--- __mocks__/openmrs-esm-api.mock.tsx
@@ -1,3 +1,0 @@
-export function openmrsFetch() {
-  return new Promise(() => {});
-}
