+++ __mocks__/openmrs-esm-framework.mock.tsx
@@ -1,4 +1,31 @@
+import { never } from "rxjs";
 import React from "react";
+
+export function openmrsFetch() {
+  return new Promise(() => {});
+}
+
+export const defineConfigSchema = jest.fn();
+
+export const validators = {
+  isUrlWithTemplateParameters: jest.fn()
+};
+
+export const interpolateString = jest.fn().mockImplementation(s => s);
+
+export const Type = {
+  Array: "Array",
+  Boolean: "Boolean",
+  ConceptUuid: "ConceptUuid",
+  Number: "Number",
+  Object: "Object",
+  String: "String",
+  UUID: "UUID"
+};
+
+export function createErrorHandler() {
+  return jest.fn().mockReturnValue(never());
+}
 
 export const openmrsComponentDecorator = jest
   .fn()
