+++ jest.config.js
@@ -9,15 +9,9 @@ module.exports = {
     "^carbon-components-react/es/(.*)$": "carbon-components-react/lib/$1",
     "lodash-es/debounce": "<rootDir>/__mocks__/debounce.mock.ts",
     "lodash-es/isEmpty": "<rootDir>/__mocks__/isEmpty.mock.ts",
-    "@openmrs/esm-api": "<rootDir>/__mocks__/openmrs-esm-api.mock.tsx",
-    "@openmrs/esm-error-handling":
-      "<rootDir>/__mocks__/openmrs-esm-error-handling.mock.tsx",
-    "@openmrs/esm-config": "<rootDir>/__mocks__/openmrs-esm-config.mock.tsx",
-    "@openmrs/esm-react-utils":
-      "<rootDir>/__mocks__/openmrs-esm-react-utils.mock.tsx",
-    "react-i18next": "<rootDir>/__mocks__/react-i18next.mock.js",
-    "@openmrs/esm-extensions":
-      "<rootDir>/__mocks__/openmrs-esm-extensions.mock.tsx"
+    "@openmrs/esm-framework":
+      "<rootDir>/__mocks__/openmrs-esm-framework.mock.tsx",
+    "react-i18next": "<rootDir>/__mocks__/react-i18next.mock.js"
   },
   setupFilesAfterEnv: ["<rootDir>/src/setupTests.ts"]
 };
